Split ear model into two classes.
The PeaqEarModel contains all static information, most of which was in
PeaqEarModelClass previously. To enable different frequency resolution
(for the advanced model) in the future, this data has been made instance
data. Objects of the new class PeaqEar contain only instance data for
paritcual instance of the ear model and a reference to a PeaqEarModel
for the static pre-computed data.